reorganize/break RequestParser in more files
