+++ webserver/Makefile
@@ -22,7 +22,8 @@ WEBSERV =		WebServ.cpp
 SERVER =		Server.cpp ServerParser.cpp server_builder.cpp server_getters.cpp \
 				server_setters.cpp
 REQUEST =		Request.cpp RequestBuilder.cpp RequestParser.cpp \
-				request_parser_getters.cpp
+				request_parser_getters.cpp parse_control_data.cpp \
+				parse_headers.cpp parse_body.cpp parse_utils.cpp
 RESPONSE =		Response.cpp ResponseBuilder.cpp StatusMessage.cpp HandleCGI.cpp \
 				HandleGetRequest.cpp HandlePostRequest.cpp HandleDeleteRequest.cpp \
 				ResponseUtils.cpp
