+++ webserver/src/Request/RequestParser.cpp
@@ -6,7 +6,7 @@
 /*   By: sguilher <sguilher@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/04 21:21:48 by sguilher          #+#    #+#             */
-/*   Updated: 2023/12/07 23:38:15 by sguilher         ###   ########.fr       */
+/*   Updated: 2023/12/08 00:35:32 by sguilher         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -97,155 +97,6 @@ void RequestParser::init(char c) {
 	method(c);
 }
 
-// general:
-// Parts of request line is separated by a space character.
-// Technically there should be only one space though I've seen badly malformed requests that send multiple spaces.
-// Browsers will never send more than one space
-// nginx aceita vários espaços e nós tb estamos aceitando
-// não aceita espaço no início da primeira linha (não é letra maiúscula)
-
-// The three fields in the initial message line should be separated by a single space, but might instead use several spaces, or tabs. Accept any number of spaces or tabs between these fields.
-
-// Method names are always uppercase
-// The method token is case-sensitive
-// When a request method is received that is unrecognized or not implemented by an origin
-// server, the origin server SHOULD respond with the 501 (Not Implemented) status code.
-// When a request method is received that is known by an origin
-// server but not allowed for the target resource, the origin server
-// SHOULD respond with the 405 (Method Not Allowed) status code.
-void RequestParser::method(char c) {
-	if (utils::is_ualpha(c))
-		_method.push_back(c);
-	else if (c == SP) {
-		if (_method.size() == 0)
-			_bad_request("invalid method: method token begins with space");
-		_step = URI;
-		Logger::debug("method", _method);
-	}
-	else if (c == HTAB)
-		_bad_request("invalid method/uri separator: horizontal tab");
-	else if (c == CR || c == LF)
-		_bad_request("missing data on request first line");
-	else if (utils::is_lalpha(c))
-		_bad_request("method characters must be upper case");
-	else
-		_bad_request("invalid character at request first line");
-}
-
-// URI enconding (tb serve pra POST) - não entendi se posso receber https
-// URI normalization:
-// normalizing the scheme and host to lowercase
-// normalizing the port to remove any leading zeros.
-// If port is elided from the URI, the default port for that scheme is used
-// o host pode ser um IP, que pode estar em ipv4 ou ipv6...
-// It is RECOMMENDED that all senders and recipients support, at a minimum, URIs with lengths of 8000 octets in protocol elements.
-void RequestParser::uri(char c) {
-	static bool init_uri = true;
-
-	if (c == SP && init_uri)
-		return ;
-	else if (c == CR || c == LF)
-		_bad_request("missing data on request first line");
-	else if (init_uri && c == HTAB)
-		_bad_request("invalid method/uri separator: horizontal tab");
-	if (http::is_uri_char(c)) {
-		if (init_uri)
-			init_uri = false;
-		_uri.push_back(c);
-	}
-	else if (c == SP) {
-		// Logger::debug("uri", _uri);
-		Logger::warning("uri", _uri);
-		_step = PROTOCOL;
-		init_uri = true;
-	}
-	else
-		_invalid_request("invalid uri character", c, http::BAD_REQUEST);
-}
-
-// The HTTP version always takes the form "HTTP/x.x", uppercase
-void RequestParser::protocol(char c) {
-	static bool init_protocol = true;
-	// static std::string protocol; // testar se dá pra usar
-
-	if (c == SP && init_protocol)
-		return ;
-	if (c != SLASH && c != CR && c != LF && c != SP) {
-		if (init_protocol)
-			init_protocol = false;
-		_protocol.push_back(c);
-	}
-	else if (c == SLASH) {
-		if (_protocol.compare(_right_protocol) != 0)
-			_invalid_request("invalid protocol", _protocol, http::BAD_REQUEST);
-			// throw http::InvalidRequest(http::NOT_FOUND); // nginx
-		_step = VERSION;
-		Logger::debug("protocol", _protocol);
-		init_protocol = true;
-	}
-	else if (c == SP)
-		_bad_request("invalid space in HTTP protocol");
-	else if (_protocol.compare(_right_protocol) == 0)
-		_bad_request("missing HTTP version");
-	else
-		_invalid_request("invalid protocol", _protocol, http::BAD_REQUEST);
-}
-
-// RFC 9110 - item 2.5. Protocol Version:
-// HTTP's version number consists of two decimal digits separated by a "."
-// (period or decimal point)
-// version format: <major-version>.<minor-version>
-
-// RFC 9110 - item 6.2. Control Data:
-// A server can send a 505 (HTTP Version Not Supported) response if it wishes,
-// for any reason, to refuse service of the client's major protocol version.
-// if major-version != 1: refuse -> 505
-
-// RFC 9110 - item 6.2. Control Data:
-// A recipient that receives a message with a major version number that it
-// implements and a minor version number higher than what it implements SHOULD
-// process the message as if it were in the highest minor version within that
-// major version to which the recipient is conformant
-// temos que aceitar 1.2, 1.3, 1.11
-// não precisamos aceitar 1.0?
-
-void RequestParser::version(char c) {
-	int size;
-
-	if (c == SP && !_version.empty())
-		return ;
-	if (std::isdigit(c) || (c == POINT && !_version.empty()))
-		_version.push_back(c);
-	else if (c == CR || c == LF) {
-		size = _version.size();
-		if (size == 3) {
-			if (_version[0] == '1' && _version[1] == POINT && _version[2] == '1') {
-				Logger::debug("version", _version);
-				if (c == CR)
-					_step = CR_FIRST_LINE;
-				if (c == LF) {
-					Logger::debug("LF first line");
-					_step = HEADER;
-				}
-			}
-			else
-				_invalid_request(
-					HTTP_VERSION, _version, http::HTTP_VERSION_NOT_SUPPORTED
-				);
-		}
-		else if (
-			size == 0
-			|| (size == 1 && _version[0] == '1')
-			|| (size == 2 && _version[0] == '1' && _version[1] == POINT)
-		)
-			_invalid_request(HTTP_VERSION, _version, http::BAD_REQUEST);
-		else
-			_invalid_request(HTTP_VERSION, _version, http::HTTP_VERSION_NOT_SUPPORTED);
-	}
-	else
-		_invalid_request(HTTP_VERSION, http::BAD_REQUEST);
-}
-
 // Each part of a HTTP request is separated by a new line
 // Note: Technically they should be \r\n but you are strongly encouraged to also accept \n as a newline
 // Initial lines and headers should end in CRLF, though you should gracefully handle lines ending in just LF.
@@ -295,131 +146,6 @@ void RequestParser::check_crlf(char c) {
 	}
 }
 
-// headers
-// servers should treat headers as an unordered set
-// one line per header, of the form "Header-Name: value", ending with CRLF
-// you should handle LF correctly
-// HTTP 1.1 defines 46 headers, and one (Host:) is required in requests.
-
-// A server MUST NOT apply a request to the target resource until it receives the
-// entire request header section, since later header field lines might include
-// conditionals, authentication credentials, or deliberately misleading duplicate
-// header fields that could impact request processing
-void RequestParser::header(char c) {
-	if (_step == HEADER) {
-		if (c == CR) {
-			_step = SECOND_CR_HEADER;
-			return ;
-		}
-		else if (c == LF) {
-			Logger::debug("LF end header");
-			_step = END;
-			return ;
-		}
-		else if (c == SP || c == HTAB) {
-			_field_name = _last_header;
-			_step = HEADER_VALUE_INIT;
-		}
-		else
-			_step = HEADER_NAME;
-	}
-	switch (_step) {
-		case HEADER_NAME:
-			_parse_field_name(c);
-			break;
-		case HEADER_VALUE_INIT:
-			if (c == SP || c == HTAB)
-				return ;
-			_step = HEADER_VALUE;
-			_parse_field_value(c);
-			break;
-		case HEADER_VALUE:
-			_parse_field_value(c);
-			break;
-		default:
-			break;
-	}
-}
-
-// Field name:
-// The requested field name. It MUST conform to the field-name syntax defined
-// in Section 5.1, and it SHOULD be restricted to just letters, digits,
-// and hyphen ('-') characters, with the first character being a letter.
-// nginx não segue essas regras, aceita qualquer coisa, google tb...
-
-// The header name is not case-sensitive (though the value may be) - Header
-// name can be either title-case or lowercase or mixed, all are valid
-
-// nos headers parece considerar um espaço no começo do nome do header como parte dele
-// interpreta " Host" e não "Host" (dá erro porque considera que o header não está incluso)
-// Header lines beginning with space or tab are actually part of the previous header line, folded into multiple lines for easy reading.
-// the following two headers are equivalent:
-
-// Header1: some-long-value-1a, some-long-value-1b
-// HEADER1:    some-long-value-1a,
-//             some-long-value-1b
-
-void RequestParser::_parse_field_name(char c) {
-	if (c != COLON && !utils::is_ctl(c) && c != SP)
-		_field_name.push_back(c);
-	else if (c == COLON) {
-		Logger::debug("key", _field_name);
-		_step = HEADER_VALUE_INIT;
-	}
-	else if (c == CR) {
-		// Logger::warning("field name without value: " + _field_name);
-		_step = CR_HEADER;
-		_add_header();
-	}
-	else
-		_invalid_request("field name character", c, http::BAD_REQUEST);
-}
-
-// Any number of spaces or tabs may be between the ":" and the value
-void RequestParser::_parse_field_value(char c) {
-	// if (c == SP || c == HTAB) // só pode ser no começo
-	// 	return ;
-	if (c == CR || c == LF) {
-		Logger::debug("value", _field_value);
-		if (c == CR)
-			_step = CR_HEADER;
-		else if (c == LF) {
-			Logger::debug("LF header");
-			_step = HEADER;
-		}
-		_add_header();
-	}
-	else if (utils::is_ctl(c) && c != HTAB)
-		_bad_request("request parser: control character on field value");
-	else
-		_field_value.push_back(c);
-}
-
-void RequestParser::_add_header(void) {
-	t_header_iterator it;
-
-	std::transform(
-		_field_name.begin(),
-		_field_name.end(),
-		_field_name.begin(),
-		&utils::c_tolower
-	);
-	_last_header = _field_name; // verificar se faz deep copy...
-	it = _headers.find(_field_name);
-	if (it == _headers.end()) {
-		// não tem o header
-		std::vector<std::string> v;
-
-		v.clear();
-		v.push_back(_field_value);
-		_headers.insert(std::make_pair(_field_name, v));
-	}
-	else // tem o header - se a linha começou com espaço/tab; se não dá erro!!
-		it->second.push_back(_field_value); // assumindo que ele já estará limpo de espaços e tabs no início e fim
-	_field_value.clear();
-	_field_name.clear();
-}
-
 void RequestParser::check_request(void) {
 	Logger::debug("Checking request");
 	// first check headers and get necessary values
@@ -438,461 +164,6 @@ void RequestParser::check_request(void) {
 	Logger::debug("End checking request");
 }
 
-// header Host is mandatory and singleton
-void RequestParser::_check_host(void) {
-	t_header_iterator it;
-
-	it = _headers.find("host");
-	if (it == _headers.end())
-		_bad_request("'Host' header not found");
-	if (it->second.size() > 1 || it->second[0].size() == 0)
-		_bad_request("'Host' header");
-	_request->setHost(it->second[0]);
-}
-
-// header Content-Length has only numbers, is singleton and has a maximum size
-void RequestParser::_check_content_length(void) {
-	t_header_iterator it_header;
-
-	it_header = _headers.find("content-length");
-	if (it_header == _headers.end())
-		return ;
-	if (it_header->second.size() > 1)
-		_bad_request("'Content-Length' header");
-	_has_content_length = true;
-
-	std::string content_lenght = it_header->second[0];
-	std::string::iterator it, end = content_lenght.end();
-	for (it = content_lenght.begin(); it != end; ++it) {
-		if (!std::isdigit(*it))
-			_invalid_request(
-				"'Content-Length' header invalid character",
-				*it,
-				http::BAD_REQUEST
-			);
-	}
-
-	std::stringstream ss(content_lenght);
-	ss >> _content_length;
-	if (_content_length > _max_body_size)
-		_invalid_request(
-			"Content-Lenght bigger than max body size",
-			http::CONTENT_TOO_LARGE
-		);
-	_request->setContentLength(_content_length);
-}
-
-void RequestParser::_check_transfer_encoding(void) {
-	t_header_iterator it_header;
-
-	it_header = _headers.find("transfer-encoding");
-	if (it_header == _headers.end())
-		return ;
-	if (it_header->second.size() > 1)
-		_bad_request("'Transfer-Encoding' header");
-	if (it_header->second[0].compare("chunked") != 0)
-		_invalid_request("Transfer-Encoding type", http::NOT_IMPLEMENTED);
-	_is_chunked = true;
-	_request->setChuncked(true);
-	Logger::debug("found Transfer-Encoding: chunked");
-}
-
-void RequestParser::_check_uri(void) {
-	// check if the path exists and set it; get location
-	// uri
-	// not found here
-	// partes da uri:
-	// scheme: não vem
-	// authority: não vem
-
-// 	The generic syntax uses the slash ("/"), question mark ("?"), and
-//    number sign ("#") characters to delimit components that are
-//    significant to the generic parser's hierarchical interpretation of an
-//    identifier
-
-// se % -> caracter especial com valor em hexadecimal % + 2 caracteres hexadecimais
-// For example, "%20" = SP
-// "%" HEXDIG HEXDIG
-// uppercase e lowercase HEXDIG são iguais
-// For consistency, URI producers and
-//    normalizers should use uppercase hexadecimal digits for all percent-
-//    encodings.
-
-// Reserved Characters:
-// reserved    = gen-delims / sub-delims
-// gen-delims  = ":" / "/" / "?" / "#" / "[" / "]" / "@" -> delimiting
-//    characters that are distinguishable from other data within a URI
-// sub-delims  = "!" / "$" / "&" / "'" / "(" / ")"
-//                   / "*" / "+" / "," / ";" / "="
-
-//  Unreserved Characters: Characters that are allowed in a URI but do not have a reserved purpose:
-// ALPHA / DIGIT / "-" / "." / "_" / "~"
-
-	// se não começa com / mandar um erro -> por hora não vou verificar scheme e authority
-	if (_uri[0] != SLASH) // localhost:8080/data
-		_invalid_request(
-			"uri different from expected", _uri, http::INTERNAL_SERVER_ERROR
-		); // ver se pode vir sem / no início
-	std::vector<t_location> locations = _request->server()->getLocations();
-	std::string::iterator it = _uri.begin(), end = _uri.end();
-	int i, location_pos = 0, locations_size = locations.size();
-	std::string path;
-
-	// get first location:
-	std::cout << BLUE << "Checking uri:\n" << GREY;
-	path.clear();
-	path.push_back(_uri[0]);
-	// /data -> /data/index.html -> passar pelos locations do server, compara com o location
-	// /data/index2.html -> /data/index2.html
-	for (i = 0; i < locations_size; ++i) {
-		if (path.compare(locations[i].location) == 0) {
-			Logger::debug("location found", locations[i].location);
-			location_pos = i;
-			break;
-		}
-	}
-	while (++it != end && !http::uri_path_end(*it)) {
-		std::cout << "char path: " << *it << std::endl;
-		if (*it == SLASH || (it + 1) == end) {
-			std::cout << "char path: " << *it << std::endl;
-			for (i = 0; i < locations_size; ++i) {
-				if (path.compare(locations[i].location) == 0) {
-					Logger::debug("location found", locations[i].location);
-					location_pos = i;
-					break;
-				}
-			}
-		}
-		path.push_back(*it); // check position
-	}
-	std::string location = locations[location_pos].location;
-	std::string root = locations[location_pos].root;
-	Logger::debug("Final path", path);
-	Logger::debug("Final location", location);
-
-	// substitui o location na path pelo root do location
-	if (location.size() > 1)
-		path.erase(0, location.size());
-	Logger::debug("Final path", path);
-	path.insert(path.begin(), root.begin(), root.end());
-	Logger::debug("Final path", path);
-
-	// se não tiver ponto => é um diretório => verificar se existe; => not found
-	//                                 => se autoindex -> cgi?
-	//                                    else index?
-	if (path.find('.') == std::string::npos) { // tem ponto na path?
-		DIR *dr;
-
-		dr = opendir(path.c_str());
-		if (dr == NULL)
-			_invalid_request("Directory not found", path, http::NOT_FOUND);
-		closedir(dr);
-		Logger::debug("Found directory", path);
-		if (!locations[location_pos].permit.autoindex) {
-			Logger::warning("Index file", *(locations[location_pos].index.begin()));
-			path = *(locations[location_pos].index.begin());
-		}
-		else {
-			// adicionar um barra no final (se não tiver)
-		}
-	}
-	else { // => é um arquivo => verificar se arquivo existe => not found
-		std::ifstream file;
-
-		file.open(path.c_str());
-		if (file.fail())
-			_invalid_request("Directory not found", path, http::NOT_FOUND);
-		file.close();
-	}
-
-	// substituir o location pelo caminho: content/(root)/arquivo?
-	// retornar o index -> /data/index.html, /data/index2.html -> iterar pra encontrar
-	// filePath
-	// localhost:8080/data/outro_caminho
-	_request->setPath(path);
-}
-
-void RequestParser::_check_method(void) {
-	// verify if method is allowed in specified location:
-	try {
-		// esse check vai ter que levar em conta o location
-		_request->setMethod(http::str_to_enum_method(_method));
-	} catch(const utils::GeneralException& e) {
-		throw http::InvalidRequest(http::METHOD_NOT_ALLOWED);
-	} catch(const std::exception& e) {
-		throw e;
-	}
-}
-
-void RequestParser::_check_post_headers(void) {
-	if (!_is_chunked && (!_has_content_length || !_content_length))
-		_invalid_request(
-			"POST without 'Transfer-Encoding' or 'Content-Length' headers",
-			http::LENGTH_REQUIRED
-		);
-	if (_is_chunked && _has_content_length && _content_length)
-		_bad_request(
-			"chunked data and content-length both setted"
-		);
-	_step = BODY;
-	if (_is_chunked)
-		_step = CHUNK_SIZE;
-}
-
-void RequestParser::_print_headers(void) {
-	if (DEBUG) {
-		t_header_iterator it, end;
-		std::vector<std::string>::iterator v_it, v_end;
-		int i;
-
-		end = _headers.end();
-		std::cout << GREY
-				<< "------------------------------------------------------\n"
-				<< BLUE << "Headers saved:\n" << GREY;
-		for (it = _headers.begin(); it != end; ++it) {
-			i = 0;
-			std::cout << it->first << ": ";
-			v_end = it->second.end();
-			for (v_it = it->second.begin(); v_it != v_end; ++v_it) {
-				++i;
-				if (i == 1)
-					std::cout << *v_it;
-				else
-					std::cout << "," << *v_it;
-			}
-			std::cout << std::endl;
-		}
-		std::cout << "------------------------------------------------------\n"
-				<< RESET;
-	}
-}
-
-// t_string_map	RequestParser::get_result(void) const {
-// 	return _result;  // tem problema? não vou alterar, só consultar...
-// }
-
-// Body
-
-//  Request messages are never close-delimited because they are always explicitly
-//  framed by length or transfer coding, with the absence of both implying the
-//   request ends immediately after the header section
-
-// A server MAY reject a request that contains a message body but not a
-// Content-Length by responding with 411 (Length Required).
-
-// If an HTTP message includes a body, there are usually header lines in the message that describe the body, eg (ver se são obrigatórios):
-// Content-Type: header gives the MIME-type of the data in the body, such as text/html or image/gif.
-// Content-Length: header gives the number of bytes in the body.
-
-// A HTTP request is terminated by two newlines
-// Note: Technically they should be 4 bytes: \r\n\r\n but you are strongly encouraged to also accept 2 byte terminator: \n\n.
-
-// If a valid Content-Length header field is present without Transfer-Encoding,
-// its decimal value defines the expected message body length in octets. If the
-// sender closes the connection or the recipient times out before the indicated
-// number of octets are received, the recipient MUST consider the message to be
-// incomplete and close the connection.
-
-void RequestParser::body(char c) {
-	if (_is_chunked)
-		return _body_chunked(c);
-	if (_body_bytes_readed < _content_length) {
-		_body.push_back(c);
-		++_body_bytes_readed;
-		if (_body_bytes_readed == _content_length) {
-			_step = BODY_LENGTH_END;
-			_print_body();
-		}
-	}
-}
-
-void RequestParser::end_body(char c) {
-	if (c == CR)
-		_step = CR_BODY;
-	else if (c == LF) {
-		Logger::debug("LF end Body");
-		_step = END;
-	}
-	else
-		_invalid_request("Body larger than specified", http::CONTENT_TOO_LARGE);
-}
-
-// RFC 9112 - 6.3.4
-// If a Transfer-Encoding header field is present in a request and the chunked
-// transfer coding is not the final encoding (0), the message body length cannot
-// be determined reliably; the server MUST respond with the 400 (Bad Request)
-//  status code and then close the connection.
-
-// depois de ler cada chunck, devolve uma resposta com status code = 202 pra
-// indicar pro cliente que foi aceito
-// quando finalizar, retorna 200 ou 201 (CREATED)
-
-// Example:
-// 4␍␊            (chunk size is four octets)
-// Wiki           (four octets of data)
-// ␍␊             (end of chunk)
-
-// 7␍␊            (chunk size is seven octets)
-// pedia i        (seven octets of data)
-// ␍␊             (end of chunk)
-
-// B␍␊            (chunk size is eleven octets)
-// n ␍␊chunks.    (eleven octets of data)
-// ␍␊             (end of chunk)
-
-// 0␍␊            (chunk size is zero octets, no more chunks)
-// ␍␊             (end of final chunk with zero data octets)
-
-// Result string: "Wikipedia in ␍␊chunks."
-// Result print:
-// "Wikipedia in
-// chunks."
-
-void RequestParser::_body_chunked(char c) {
-	switch (_step) {
-		case CHUNK_SIZE:
-			_parse_chunk_size(c);
-			break;
-		case CHUNK_PARAMETERS:
-			// por hora não vou fazer nada com os parametros
-			if (c == CR)
-				_step = CR_CHUNK_SIZE;
-			if (c == LF)
-				_step = CHUNK_DATA;
-			Logger::debug("chunk parameter: ", c);
-			break;
-		case CHUNK_DATA:
-			_parse_chunk_data(c);
-			break;
-		case CHUNK_DATA_END:
-			if (c == CR)
-				_step = CR_CHUNK_DATA;
-			else if (c == LF)
-				_step = CHUNK_SIZE;
-			else // verificar o tipo de erro
-				_invalid_request(
-					"chunk data larger than specified", http::CONTENT_TOO_LARGE
-				);
-			break;
-		case CHUNK_END:
-			if (c == CR)
-				_step = SECOND_CR_CHUNK_END;
-			else if (c == LF)
-				_step = END;
-			else // verificar o tipo de erro
-				_invalid_request(
-					"chunk data larger than specified", http::CONTENT_TOO_LARGE
-				);
-			break;
-
-		default:
-			break;
-	}
-	// _print_body();
-}
-
-void RequestParser::_parse_chunk_size(char c) {
-	if (c == CR || c == LF || c == SEMICOLON) {
-		if (c == CR)
-			_step = CR_CHUNK_SIZE;
-		else if (c == LF)
-			_step = CHUNK_DATA;
-		else if (c == SEMICOLON)
-			_step = CHUNK_PARAMETERS;
-		_chunk_size = utils::xtod(_chunk_size_str);
-		_chunk_bytes_readed = 0;
-		Logger::debug("chunk size hex:", _chunk_size_str);
-		Logger::debug("chunk size decimal:", _chunk_size);
-		if (_chunk_size == 0) {
-			if (c == CR)
-				_step = CR_CHUNK_END;
-			if (c == LF) {
-				Logger::debug("first LF chunk end");
-				_step = CHUNK_END;
-			}
-		}
-		_chunk_size_str.clear();
-	}
-	else if (!std::isxdigit(c))
-		_bad_request("Invalid character on chunk size");
-	else
-		_chunk_size_str.push_back(c);
-}
-
-void RequestParser::_parse_chunk_data(char c) {
-	if (_chunk_bytes_readed < _chunk_size) {
-		_body.push_back(c);
-		++_chunk_bytes_readed;
-		if (_chunk_bytes_readed == _chunk_size) {
-			_step = CHUNK_DATA_END;
-			_print_body();
-		}
-	}
-}
-
-
-void RequestParser::_print_body(void) {
-	std::vector<char>::iterator it, end = _body.end();
-
-	if (DEBUG) {
-		std::cout << GREY
-				<< "------------------------------------------------------\n"
-				<< BLUE << "Received body of size " << _body_bytes_readed
-				<< ":\n" << GREY;
-		for(it = _body.begin(); it != end; ++it) {
-			// if (*it == CR)
-			// 	std::cout << " CR";
-			// else if (*it == LF)
-			// 	std::cout << "LF" << *it;
-			// else
-			std::cout << *it;
-		}
-		std::cout
-				<< "\n------------------------------------------------------\n"
-				<< RESET << std::endl;
-	}
-}
-
-void RequestParser::_bad_request(std::string const description) {
-	http::bad_request(REQPARSER_ERROR + description);
-}
-
-void  RequestParser::_invalid_request(
-	std::string const description,
-	std::string const value,
-	http::HttpStatus error_code
-) {
-	Logger::warning(REQPARSER_ERROR + description, value);
-	throw http::InvalidRequest(error_code);
-}
-
-void  RequestParser::_invalid_request(
-	std::string const description,
-	char const c,
-	http::HttpStatus error_code
-) {
-	Logger::warning_no_lf(REQPARSER_ERROR + description);
-	std::cout << GREY << c << std::endl;
-	throw http::InvalidRequest(error_code);
-}
-
-void  RequestParser::_invalid_request(
-	std::string const description,
-	http::HttpStatus error_code
-) {
-	Logger::warning(REQPARSER_ERROR + description);
-	throw http::InvalidRequest(error_code);
-}
-
-// void RequestParser::break_data(char* buffer, size_t bytes_read) {
-// 	for (size_t i = 0; i < bytes_read; ++i)
-// 		_body.push_back(buffer[i]);
-
-// 	if (DEBUG) {
-// 		std::cout << GREY << "Received/remaining data:\n";
-// 		_print_body();
-// 	}
-// }
 
 // The most common use of POST, by far, is to submit HTML form data to CGI scripts.
 // In this case, the Content-Type: header is usually application/x-www-form-urlencoded,
